In eo:, edit pages are in x-convention coding, but XML export uses UTF. Specify that sometimes 'x' has to be changed into 'xx'.
Unfortunately, I have not been able to also change the UTF-8 entities into x-convention, which is no problem on eo: itself, but causes us to not find links to other languages when the titel contains 'ux' (or one of the other combinations of the x-conversion). Could someone else please look into that? Thanks in advance.